+++ packages/i18n/src/parser/properties/constant.ts
@@ -1,0 +1,4 @@
+export const SEPARATOR = /[:=]/;
+export const COMMENT_START = /[#!]/;
+export const END_OF_LINE = /\r|\n|\r\n/;
+export const WHITESPACE = /[ \t\f]+/;