+++ packages/project-access/src/project/index.ts
@@ -11,9 +11,10 @@ export {
     readCapServiceMetadataEdmx
 } from './cap';
 export { getNodeModulesPath } from './dependencies';
-export { getI18nPropertiesPaths } from './i18n';
+export { getCapI18nFolderNames } from './i18n';
 export { getAppProgrammingLanguage, getAppType, getProject, getProjectType } from './info';
 export { loadModuleFromProject } from './module-loader';
 export { findAllApps, findCapProjects, findFioriArtifacts, findProjectRoot, getAppRootFromWebappPath } from './search';
 export { getWebappPath, readUi5Yaml } from './ui5-config';
 export { getMtaPath } from './mta';
+export { createApplicationAccess, createProjectAccess } from './access';