+++ packages/create/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ux/create
 
+## 0.5.28
+
+### Patch Changes
+
+-   Updated dependencies [86da13e3]
+    -   @sap-ux/project-access@1.17.1
+    -   @sap-ux/adp-tooling@0.8.5
+    -   @sap-ux/app-config-writer@0.3.23
+    -   @sap-ux/cap-config-writer@0.2.12
+    -   @sap-ux/mockserver-config-writer@0.2.2
+
 ## 0.5.27
 
 ### Patch Changes