+++ packages/ui5-library-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/ui5-library-writer
 
+## 0.3.12
+
+### Patch Changes
+
+-   Updated dependencies [86da13e3]
+    -   @sap-ux/project-access@1.17.1
+    -   @sap-ux/ui5-application-writer@0.26.2
+
 ## 0.3.11
 
 ### Patch Changes